Delete DistantCopCarSirens.md
The sounds are very bugging to mostly everyone playing FiveM. I know for me, I get on to relax but I just can't when that sounds is CONSTANTLY playing in my ear. This is not to talk down on it or whoever added it in, I would gladly appreciate it if this was considered and it was just taken out so we can enjoy the game.